+++ src/components/App/App.css
@@ -2,8 +2,8 @@
     font-family: 'Inter', 'Arial', sans-serif;
     display: flex;
     flex-direction: column;
-    justify-content: center;
     max-width: 1280px;
     min-width: 320px;
     margin: auto;
+    min-height: 100%;
 }
