+++ src/components/App/AppRouter/AppRouter.jsx
@@ -6,17 +6,21 @@ import Login from '../../Login/Login';
 import Profile from '../../Profile/Profile';
 import SavedMovies from '../../SavedMovies/SavedMovies';
 import Movies from '../../Movies/Movies';
+import Layout from '../Layout/Layout';
 
 class AppRouter extends Component {
     render() {
         return (
             <Switch>
-                <Route exact path='/' children={<Main/>}/>
-                <Route exact path='/movies' children={<Movies/>}/>
-                <Route exact path='/saved-movies' children={<SavedMovies/>}/>
-                <Route exact path='/profile' children={<Profile/>}/>
                 <Route exact path='/signin' children={<Login/>}/>
                 <Route exact path='/signup' children={<Register/>}/>
+
+                <Layout>
+                    <Route exact path='/' children={<Main/>}/>
+                    <Route exact path='/movies' children={<Movies/>}/>
+                    <Route exact path='/saved-movies' children={<SavedMovies/>}/>
+                    <Route exact path='/profile' children={<Profile/>}/>
+                </Layout>
             </Switch>
         );
     }
